requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==15.3.1
install==1.3.5
Pillow==9.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
